adding Redis cache for clear PENDING status when not payment over 15min, add ing Booking business

+++ src/main/java/com/example/bookingapp/repository/RoomRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
     List<Room> findByPropertyId(Long propertyId);
 }
